pyproject.toml
[tool.poetry]
name = "percepta-inference"
version = "0.0.1"
description = ""
authors = ["Noel Gomariz"]
readme = "README.md"

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[tool.poetry.dependencies]
python = "^3.12"
python-dotenv = "1.1.0"
requests = "2.32.3"
numpy = "2.1.1"
toml = "^0.10.2"
opencv-python = "^4.11.0.86"
pillow = "11.1.0"
redis = "5.2.1"
hiredis = "3.1.0"
ultralytics = "8.3.110"
